+++ tcod/libtcodpy.py
@@ -1240,7 +1240,10 @@ def path_new_using_function(w, h, func, userData=0, dcost=1.41):
     Returns:
         AStar: A new AStar instance.
     """
-    return tcod.path.AStar((func, userData), dcost, w, h)
+    return tcod.path.AStar(
+        tcod.path._EdgeCostFunc((func, userData), w, h),
+        dcost,
+        )
 
 def path_compute(p, ox, oy, dx, dy):
     """Find a path from (ox, oy) to (dx, dy).  Return True if path is found.
@@ -1254,7 +1257,7 @@ def path_compute(p, ox, oy, dx, dy):
     Returns:
         bool: True if a valid path was found.  Otherwise False.
     """
-    return lib.TCOD_path_compute(p.cdata, ox, oy, dx, dy)
+    return lib.TCOD_path_compute(p._path_c, ox, oy, dx, dy)
 
 def path_get_origin(p):
     """Get the current origin position.
@@ -1268,7 +1271,7 @@ def path_get_origin(p):
     """
     x = ffi.new('int *')
     y = ffi.new('int *')
-    lib.TCOD_path_get_origin(p.cdata, x, y)
+    lib.TCOD_path_get_origin(p._path_c, x, y)
     return x[0], y[0]
 
 def path_get_destination(p):
@@ -1281,7 +1284,7 @@ def path_get_destination(p):
     """
     x = ffi.new('int *')
     y = ffi.new('int *')
-    lib.TCOD_path_get_destination(p.cdata, x, y)
+    lib.TCOD_path_get_destination(p._path_c, x, y)
     return x[0], y[0]
 
 def path_size(p):
@@ -1292,7 +1295,7 @@ def path_size(p):
     Returns:
         int: Length of the path.
     """
-    return lib.TCOD_path_size(p.cdata)
+    return lib.TCOD_path_size(p._path_c)
 
 def path_reverse(p):
     """Reverse the direction of a path.
@@ -1302,7 +1305,7 @@ def path_reverse(p):
     Args:
         p (AStar): An AStar instance.
     """
-    lib.TCOD_path_reverse(p.cdata)
+    lib.TCOD_path_reverse(p._path_c)
 
 def path_get(p, idx):
     """Get a point on a path.
@@ -1313,7 +1316,7 @@ def path_get(p, idx):
     """
     x = ffi.new('int *')
     y = ffi.new('int *')
-    lib.TCOD_path_get(p.cdata, idx, x, y)
+    lib.TCOD_path_get(p._path_c, idx, x, y)
     return x[0], y[0]
 
 def path_is_empty(p):
@@ -1324,7 +1327,7 @@ def path_is_empty(p):
     Returns:
         bool: True if a path is empty.  Otherwise False.
     """
-    return lib.TCOD_path_is_empty(p.cdata)
+    return lib.TCOD_path_is_empty(p._path_c)
 
 def path_walk(p, recompute):
     """Return the next (x, y) point in a path, or (None, None) if it's empty.
@@ -1341,7 +1344,7 @@ def path_walk(p, recompute):
     """
     x = ffi.new('int *')
     y = ffi.new('int *')
-    if lib.TCOD_path_walk(p.cdata, x, y, recompute):
+    if lib.TCOD_path_walk(p._path_c, x, y, recompute):
         return x[0], y[0]
     return None,None
 
@@ -1353,36 +1356,39 @@ def dijkstra_new(m, dcost=1.41):
     return tcod.path.Dijkstra(m, dcost)
 
 def dijkstra_new_using_function(w, h, func, userData=0, dcost=1.41):
-    return tcod.path.Dijkstra((func, userData), dcost, w, h)
+    return tcod.path.Dijkstra(
+        tcod.path._EdgeCostFunc((func, userData), w, h),
+        dcost,
+        )
 
 def dijkstra_compute(p, ox, oy):
-    lib.TCOD_dijkstra_compute(p.cdata, ox, oy)
+    lib.TCOD_dijkstra_compute(p._path_c, ox, oy)
 
 def dijkstra_path_set(p, x, y):
-    return lib.TCOD_dijkstra_path_set(p.cdata, x, y)
+    return lib.TCOD_dijkstra_path_set(p._path_c, x, y)
 
 def dijkstra_get_distance(p, x, y):
-    return lib.TCOD_dijkstra_get_distance(p.cdata, x, y)
+    return lib.TCOD_dijkstra_get_distance(p._path_c, x, y)
 
 def dijkstra_size(p):
-    return lib.TCOD_dijkstra_size(p.cdata)
+    return lib.TCOD_dijkstra_size(p._path_c)
 
 def dijkstra_reverse(p):
-    lib.TCOD_dijkstra_reverse(p.cdata)
+    lib.TCOD_dijkstra_reverse(p._path_c)
 
 def dijkstra_get(p, idx):
     x = ffi.new('int *')
     y = ffi.new('int *')
-    lib.TCOD_dijkstra_get(p.cdata, idx, x, y)
+    lib.TCOD_dijkstra_get(p._path_c, idx, x, y)
     return x[0], y[0]
 
 def dijkstra_is_empty(p):
-    return lib.TCOD_dijkstra_is_empty(p.cdata)
+    return lib.TCOD_dijkstra_is_empty(p._path_c)
 
 def dijkstra_path_walk(p):
     x = ffi.new('int *')
     y = ffi.new('int *')
-    if lib.TCOD_dijkstra_path_walk(p.cdata, x, y):
+    if lib.TCOD_dijkstra_path_walk(p._path_c, x, y):
         return x[0], y[0]
     return None,None